tests: add more 4xx
